requirements.txt
flask>=2.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
